Add generic ThriftResponse
